+++ addon/core/say-editor.ts
@@ -52,7 +52,6 @@ export default class SayEditor {
   @tracked mainView: SayView;
   @tracked activeView: SayView;
   @tracked showRdfaBlocks = false;
-  @tracked showFormattingMarks = false;
   owner: Owner;
   root: Element;
   baseIRI: string;